reorganized time helpers
