+++ TimeUtils.h
@@ -29,83 +29,9 @@
 
 */
 
-static bool RTC_is_running = false;
-// some date/time formats used in this app
-const char* hhmmStringTpl = "  %02d:%02d  ";
-static char hhmmString[13] = "  --:--  ";
-const char* hhmmssStringTpl = "%02d:%02d:%02d";
-static char hhmmssString[13] = "--:--:--"; 
-const char* UpTimeStringTpl = "  %02d:%02d  ";
-static char UpTimeString[13] = "  --:--  ";
-const char* YYYYMMDD_HHMMSS_Tpl = "%04d-%02d-%02d %02d:%02d:%02d";
-static char YYYYMMDD_HHMMSS_Str[32] = "YYYY-MM-DD HH:MM:SS";
-static bool dayChangeTrigger = false;
-
-int current_day = -1;
-int current_hour = -1;
-
-enum TimeUpdateSources {
-  SOURCE_NONE = 0,
-  SOURCE_COMPILER = 1,
-  SOURCE_RTC = 2,
-  SOURCE_NTP = 3,
-  SOURCE_BLE = 4
-};
-
-void logTimeActivity(TimeUpdateSources source, int epoch) {
-  preferences.begin("BLEClock", false);
-  preferences.clear();
-  //DateTime epoch = RTC.now();
-  preferences.putUInt("epoch", epoch);
-  preferences.putUChar("source", source);
-  preferences.end();
-}
-
-void resetTimeActivity(TimeUpdateSources source) {
-  preferences.begin("BLEClock", false);
-  preferences.clear();
-  preferences.putUInt("epoch", 0);
-  preferences.putUChar("source", source);
-  preferences.end();
-}
-
-struct TimeActivity {
-  DateTime epoch;
-  byte source;
-};
-
-void TimeInit() {
-  preferences.begin("BLEClock", true);
-  lastSyncDateTime = preferences.getUInt("epoch", millis());
-  byte clockUpdateSource   = preferences.getUChar("source", 0);
-  preferences.end();
-  sprintf(YYYYMMDD_HHMMSS_Str, YYYYMMDD_HHMMSS_Tpl, 
-    lastSyncDateTime.year(),
-    lastSyncDateTime.month(),
-    lastSyncDateTime.day(),
-    lastSyncDateTime.hour(),
-    lastSyncDateTime.minute(),
-    lastSyncDateTime.second()
-  );
-  log_e("Defrosted lastSyncDateTime from NVS (may be bogus) : %s - source : %d", YYYYMMDD_HHMMSS_Str, clockUpdateSource);
-  #if RTC_PROFILE > HOBO
-    if(clockUpdateSource==SOURCE_NONE) {
-      if(RTC.isrunning()) {
-        log_d("[RTC] Forcing source to RTC and rebooting");
-        logTimeActivity(SOURCE_RTC, 0);
-        ESP.restart();
-      } else {
-        log_d("[RTC] isn't running!");
-        return;
-      }
-    }
-    nowDateTime = RTC.now();
-    Time_is_set = true;
-  #endif
-}
 
 
-static void timeHousekeeping(bool checkNTP=false) {
+static void timeHousekeeping() {
   unsigned long seconds_since_boot = millis() / 1000;
   unsigned long  minutes_since_boot = seconds_since_boot / 60;
   unsigned long  mm = minutes_since_boot % 60;
@@ -113,39 +39,35 @@ static void timeHousekeeping(bool checkNTP=false) {
   unsigned long  ss = seconds_since_boot % 60;
   
   DateTime internalDateTime = DateTime(year(), month(), day(), hour(), minute(), second());
-  
-  #if RTC_PROFILE > HOBO
+
+  #if RTC_PROFILE == HOBO
+
+    if( abs( seconds_since_boot - internalDateTime.unixtime() ) > 2 ) {
+      // safe to assume internal RTC is running
+      sprintf(hhmmString, hhmmStringTpl, internalDateTime.hour(), internalDateTime.minute());
+      Time_is_set = true;
+      nowDateTime = internalDateTime;
+      if( current_hour != internalDateTime.hour() ) {
+        hourChangeTrigger = true;
+      }
+    }
+    sprintf(hhmmssString, hhmmssStringTpl, hh, mm, ss);
+
+  #else // RTC_PROFILE != HOBO
 
     #ifdef BUILD_NTPMENU_BIN // NTPMENU mode
+
       nowDateTime = internalDateTime;
+
     #else // CHRONOMANIAC mode
+
+      if( current_hour != internalDateTime.hour() ) {
+        hourChangeTrigger = true;
+      }
       // - get the time from the internal clock (saves I2C calls)
-      //DateTime internalDateTime = DateTime(year(), month(), day(), hour(), minute(), second());
       // - compare internal RTC time and external RTC time every hour
-      if( current_hour != internalDateTime.hour() || checkNTP) {
-        DateTime externalDateTime = RTC.now();
-        if( checkNTP ) {
-          int64_t seconds_since_last_ntp_update = abs( externalDateTime.unixtime() - lastSyncDateTime.unixtime() );
-          log_e("seconds_since_last_ntp_update = now(%d) - last(%d) = %d seconds", externalDateTime.unixtime(), lastSyncDateTime.unixtime(), seconds_since_last_ntp_update);
-          if ( seconds_since_last_ntp_update > 86400 ) {
-            Serial.println("Stalling");
-            rollBackOrUpdateFromFS( BLE_FS, NTP_MENU_FILENAME );
-          } else {
-            setTime( externalDateTime.unixtime() );
-          }
-        } else {
-          int64_t drift = abs( externalDateTime.unixtime() - internalDateTime.unixtime() );
-          Serial.printf("[Hourly Synching Clocks] : %d seconds drift", drift);
-          // - adjust internal RTC
-          setTime( externalDateTime.unixtime() );
-          //internalDateTime = DateTime(year(), month(), day(), hour(), minute(), second());
-          //drift = abs( internalDateTime.unixtime() - externalDateTime.unixtime() );
-          if(drift > 1) {
-            log_e("[***** WTF Clocks don't agree after adjustment] %d - %d = %d\n", internalDateTime.unixtime(), externalDateTime.unixtime(), drift);  
-          } else {
-            log_e("[***** OK Clocks agree] %d - %d = %d\n", internalDateTime.unixtime(), externalDateTime.unixtime(), drift);
-          }
-        }
+      if( hourChangeTrigger ) {
+        checkForTimeUptade( internalDateTime ); // and update internal clock if necessary
         internalDateTime = DateTime(year(), month(), day(), hour(), minute(), second());
         current_hour = internalDateTime.hour();
       }
@@ -156,32 +78,13 @@ static void timeHousekeeping(bool checkNTP=false) {
         current_day = internalDateTime.day();
       }
       nowDateTime = internalDateTime;
-      
+
     #endif // ifndef BUILD_NTPMENU_BIN
 
     sprintf(hhmmString, hhmmStringTpl, internalDateTime.hour(), internalDateTime.minute());
     sprintf(hhmmssString, hhmmssStringTpl, internalDateTime.hour(), internalDateTime.minute(), internalDateTime.second());
-    #if RTC_PROFILE == CHRONOMANIAC  // chronomaniac mode
-      /*
-      if (checkNTP && RTC.isrunning()) {
-        int64_t deltaInSeconds = abs( internalDateTime.unixtime() - lastSyncDateTime.unixtime() );
-        log_e("now(%d) - last(%d) = %d seconds", internalDateTime.unixtime(), lastSyncDateTime.unixtime(), deltaInSeconds);
-        if ( deltaInSeconds > 86400) {
-          log_e("[CHRONOMANIAC] Last Time Sync: %s (%d seconds ago). Time isn't fresh anymore, should reload NTP menu !!", YYYYMMDD_HHMMSS_Str, deltaInSeconds);
-          Serial.println("Stalling");
-          rollBackOrUpdateFromFS( BLE_FS, NTP_MENU_FILENAME );
-          ESP.restart();
-        }
-      }*/
-    #endif // RTC_PROFILE == CHRONOMANIAC
-  #else // RTC_PROFILE <= HOBO
-    if( abs( seconds_since_boot - internalDateTime.unixtime() ) > 2 ) {
-      sprintf(hhmmString, hhmmStringTpl, internalDateTime.hour(), internalDateTime.minute());
-      Time_is_set = true;
-      nowDateTime = internalDateTime;
-    }
-    sprintf(hhmmssString, hhmmssStringTpl, hh, mm, ss);
-  #endif // RTC_PROFILE > HOBO
+
+  #endif // RTC_PROFILE != HOBO
 
   sprintf(UpTimeString, UpTimeStringTpl, hh, mm);
   log_d("Time:%s, Uptime:", hhmmString, UpTimeString );
@@ -209,19 +112,11 @@ bool RTCSetup() {
 }
 
 
-
-#ifdef BUILD_NTPMENU_BIN
-  #include "NTP.h"
-#endif
-
 void timeSetup() {
   #if RTC_PROFILE==HOBO
     return;
   #else 
-    //setenv("TZ", TZ_INFO, 1);
-    //tzset();
-    if(!RTCSetup()) {
-      // RTC failure ....
+    if(!RTCSetup()) { // RTC failure ....
       log_e("RTC Failure, switching to hobo mode");
     }
     #ifdef BUILD_NTPMENU_BIN
@@ -230,5 +125,5 @@ void timeSetup() {
       TimeInit();
     #endif
   #endif
-  timeHousekeeping( true );
+  timeHousekeeping();
 }
